Move type definition into commands file
